move occ config validator and occ config from meta tag factory -> to occ module
