+++ projects/core/src/config/index.ts
@@ -1,5 +1,3 @@
 export * from './config.module';
 export * from './server-config/server-config';
-export * from './server-config/server-config-from-meta-tag-factory';
-export * from './server-config/server-config-validator';
 export * from './utils/config-validator';
