--- projects/core/src/config/server-config/server-config-validator.ts
@@ -1,7 +1,0 @@
-import { OccConfig } from '../../occ';
-
-export function serverConfigValidator(config: OccConfig) {
-  if (config.backend.occ.baseUrl === undefined) {
-    return 'Please configure server.baseUrl before using storefront library!';
-  }
-}
