+++ projects/core/src/occ/config/occ-config-validator.ts
@@ -1,0 +1,11 @@
+import { OccConfig } from '../../occ';
+
+export function occConfigValidator(config: OccConfig) {
+  if (
+    config.backend === undefined ||
+    config.backend.occ === undefined ||
+    config.backend.occ.baseUrl === undefined
+  ) {
+    return 'Please configure backend.occ.baseUrl before using storefront library!';
+  }
+}
