+++ projects/core/src/occ/index.ts
@@ -1,5 +1,7 @@
 export * from './config/occ-config';
 export * from './config/default-occ-config';
+export * from './config/occ-config-from-meta-tag-factory';
+export * from './config/occ-config-validator';
 export * from './occ.module';
 export * from './utils/interceptor-util';
 export * from './miscs/miscs.service';
